[FIX] Modified recipes to avoid blocking the flow when waiting the object name.
It seems that after getting the object_name from ASR (asked by a goal recipe) the father recipe did not enter in the next receipe waiting for object name.
This seems a bug since the atoms were received. Perhaps the problem was that the atom was received even before the receipe did know that there is another goal waiting for this atom.
The way I used to solve it is to put the object_name_teller recipe as triggerable :)